Update CONTRIBUTING.md
My intention here is to unify the format (only markdown) and fix how
the `rails c` command in the guide looked.

Also removed the 'Do you want to contribute?' text because it feels
redundant with the 'Grab an issue' text.

And added a link to the github pull request documentation.