Fix view tests for search functionality
The tests didn't test the intended behavior since profiles are displayed
on the landing page even without searching, and rspec cannot follow links
when e.g. clicking a button. So, instead, we test for presence of correct
input fields and buttons, and that correct profiles are displayed when
going directly to the search url