Fix category view of profiles
that was broken due to a huge amount of profiles that had
to pass through the profile_card_details method.
We don't need to do that here because we can pass pagy different kinds
of collections. We only need the pagy_arrays for the search results.
All other types of results can be passed as an ActiveRecordRelation.

We also refactored the profile_card_details to be a Struct. This allows
us to use the same partial to display the profile cards with the
different kinds of collections.

Co-authored-by: Andy <andreas.hellwig@bitbirds.berlin>